build.rs can be run from arbitrary directories
Currently it hard-codes an assumption that the build script is run from
the $CARGO_MANIFEST_DIR. While this is true for cargo, and typically is
true for other build systems, some toolchains may have other
requirements.

1e1021d05bdd41cb02bbc43615c9dd7ecd35f853 started encoding that this
assumption may not always hold, this commit just extends that other
places the assumption is currently made.

This was tested by adding a call to
`std::env::set_current_dir("/tmp").unwrap()` at the top of the build
script's `main` function, and has been used to build both the
pregenerated version and the non-pregenerated version of the build
script.

I agree to license my contributions to each file under the terms given
at the top of each file I changed.